requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
pdfplumber
pypdf
python-docx
docx2txt
numpy
pandas
torch --index-url https://download.pytorch.org/whl/cpu
